refactor: CardItem과 Cards 분리 및 그 외 자잘한 수정
